fix: fix issue that impeded gitcoin action to be completed

+++ apps/govquests-api/govquests/action_tracking/lib/action_tracking/strategies/gitcoin_score_action_strategy.rb
@@ -24,7 +24,7 @@ module ActionTracking
       nonce = data["nonce"]
 
       response = @gitcoin_api.submit_passport(wallet_address, signature, nonce)
-      score = response["score"].to_i
+      score = response["score"].to_f
 
       if score > GITCOIN_SCORE_HUMANITY_THRESHOLD
         {
